[WIP] Got UploadFile action to complete. commitTransaction not working
I removed and fixed some of the headers for the uploadFile action and it's completing successfully now. I had to change it to upload all of the file in one go, but I left the chunkSize variable in place in case we want to revisit it in the future. The commitTransaction function that actually saves the file is still failing. I can't even get it to succeed through Insomnia, and I'm not 100% sure what the arguments need to be.